Extract common classes to shared_utils
